feat: run until end of duration without requests count
Separate behaviors:
- sem.Run becomes sem.RunFor. It quits either when are requests done or when duration is over.
- sem.RunUntil only quits when duration is over (for ever loop).

Passing a positive requests count as a flag results in running with RunFor.
Omitting the flag results in running RunUntil.

Both functions now process the results concurrently (as opposed to after the quit signal is received).
They use an internal done channel to signal the results channel has been emptied by the goroutine.
